DWARF: Track the positions of 'end', 'else', 'catch' binary locations (#2603)
Control flow structures have those in addition to the normal span of
(start, end), and we need to track them too.

Tracking them during reading requires us to track control flow
structures while parsing, so that we can know to which structure
an end/else/catch refers to.

We track these locations using a map on the side of instruction
to its "extra" locations. That avoids increasing the size of the
tracking info for the much more common non-control flow
instructions.

Note that there is one more 'end' location, that of the function
(not referring to any instruction). I left that to a later PR to
not increase this one too much.