Fix for cmake 3.10 (eg on Ubuntu LTS)
CMAKE_PROJECT_VERSION is only predefined on cmake 3.12 and later,
so the previous code produced an empty version number which leads
to parsing errors when emcc checks the version.

Use of the older PROJECT_VERSION variable as the source of the
original version works here, as there's only one toplevel project
defined.