Disable Flatten-related passes when EH is enabled
This adds `supportedFeatures` field to `Pass`, which is set to all
features by default. Individual passes can opt to disable some of the
features in their constructor, for example, currrently Flatten and other
passes depending on Flatten cannot run with exception handling
instructions.

This also adds a flag `--force-unsupported-passes`, mainly for testing
purpose. Even though Flatten and dependent passes are not fully
supported for EH, the development is in progress and we have some tests
for that, so we need a way to run those passes in tests.