Concise error output (#2652)
Don't print the entire module on an error. Instead, just print
the validation errors.

However, if the user passed --print, then do print it, as otherwise
nothing would get printed - the error would be before the pass
to print happens. And in general a user passing in a request
to print would expect a printed module anyhow.

fixes #2634