Expose asyncify state via a getter
Normally, a wrapper has to track state separately to know when to unwind/rewind and when to actually call import functions.

Exposing Asyncify state can help avoid this duplication and avoid subtle bugs when internal and wrapper state get out of sync.

Since this is a tiny function and it's useful for any Asyncify embedder, I've decided to expose it by default rather than hide behind an option.