Emit tuples in the fuzzer
tuple.make and tuple.extract instructions are always enabled, but
control flow is only allowed to have tuple types when the multivalue
feature is enabled. Also slightly refactors the top-level
`makeConcrete` function to be more selective about what it tries to
make based on the requested type to reduce the number of trivial nodes
created because the requested type is incompatible with the requested
node.