Handle tuple locals in Asyncify call support
Instead of producing temporary getter and setter functions for a
finite set of types, collect the set of local types in the module and
produce getter and setter functions for each of them. This change
temporarily has tuple values as arguments to the temporary functions,
which is invalid, but this is not a problem in practice because those
functions are later replaced with the original local.get and local.set
instructions.