Throw exceptions for when EH is used in unsupported code
This adds `UnsupportedException` class and make code throw this when EH
instructions are used in unsupported code. I think it is better than
assertion because this is something that can actually happen, when a
user tries to run specific passes using pass arguments on wasm files
that contain EH instructions.

Currently EH instructions are unsupported in Flatten, ReReloop, and
DataFlow-using passes.