Work around a compiler issue on MacOS (#2730)
The post-commit Mac bot on github failed to compile #2727 with

Undefined symbols for architecture x86_64:
  "wasm::ReorderLocals::UNSEEN", referenced from:
      wasm::ReorderLocals::doWalkFunction(wasm::Function*) in ReorderLocals.cpp.o
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)

That seems odd, it's defined like this?

  static const Index UNSEEN = -1;

Perhaps we are hitting a weird compiler bug. Anyhow, just use an
enum to be safe.

Also for clarity switch to 0 for the unseen value.