Remove old setup.cfg
I believe this was originally used to drive the pep8 (which is now
called pycodestyle) and flake8 (which we still use and is configured
via the .flake8 file).  Its clear it hasn't been used in a while since
the warning suppressions are all wrong now (no now use 4 space indents).