Dummy interpreter support for EH
This adds dummy interpreter support for EH instructions, mainly for
fuzzing. The plan is to make the interpreter support for EH instructions
correctly using Asyncify in the future. Also to support the correct
behavior we will need a `Literal` of `exnref` type too, which will be
added later too.

Currently what this dummy implementation does is:
- `try`-`catch`-`end`: only runs `try` body and ignores `catch` body
- `throw`: traps
- `retyrow`:
  - Traps on nullref argument (correct behavior based on the spec)
  - Traps otherwise too (dummy implementation for now)
- `br_on_exn`:
  - Traps on nullref (correct behavior)
  - Otherwise we assume the current expression matches the current event
    and extracts a 0 literal based on the current type.

This also adds some interpreter tests, which tests the basic dummy
behaviors for now. (Deleted tests are the ones that weren't tested
before.)