Add stacky mode to BinaryenIR
As an alternative to StackIR, adds a "stacky" mode to the normal
BinaryenIR in which normal instructions are completely flat and have
only pops as their children. This means that all instructions in
blocks may have arbitrary concrete types, not just the last
instructions.

The potential benefits of introducing a new mode for Binaryen IR are
that many passes will be agnostic to whether they run on the stacky or
unstacky code and stacky optimizations might be easier to write. This
will require further comparative exploration.

The next steps in this direction are:

 - Perform the actual tuple lowering in Stackify.cpp

 - Add a new validation mode that checks invariants for stacky code

 - Testing and fuzzing support

 - Reimplement the StackIR optimizations on the stacky BinaryenIR

 - Add checking to the pass runner to ensure passes get only code
   forms they can handle.

No more work will be done on this PR until tuples are made to work
with the current StackIR. Then maybe we can look more concretely at
the tradeoffs involved.