Make Stack IR a canonical representation
Rather than a temporary side data structure. This means that it is now
an error to run a normal pass in the presence of Stack IR because the
normal IR may be out of date and this can no longer be solved by
throwing away the Stack IR. The only way to transition back from
having Stack IR right now is to perform a round trip, so the fuzzer
and various tools are updated to add these round trip transitions
where necessary to avoid an error.

This change is a prerequisite for being able to lower tuples away when
generating Stack IR because lowering tuples requires changing a
Function's locals, which invalidates the normal IR.