Make reduction even easier from fuzz_opt.py (#2832)
With this, when it finds a bug all you need to do is copy-paste
a single line and it runs the reducer. To do that, it creates a
reducer script and fills it out for you. The script has all the docs
we used to print out to the console, so the console output is
more focused and concise now.

In most cases just running the single line it suggests should
work, however I found that it doesn't always. One reason is
#2831

Also add a missing sys.exit(1), without which we returned 0
despite printing out an error on a failing testcase.